module: Add retpoline tag to VERMAGIC
 Add a marker for retpoline to the module VERMAGIC. This catches the
case when a non RETPOLINE compiled module gets loaded into a retpoline
kernel, making it insecure.

It doesn't handle the case when retpoline has been runtime disabled.
Even in this case the match of the retcompile status will be enforced.
This implies that even with retpoline run time disabled all modules
loaded need to be recompiled.

Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Reviewed-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Acked-by: David Woodhouse <dwmw@amazon.co.uk>
Cc: rusty@rustcorp.com.au
Cc: arjan.van.de.ven@intel.com
Cc: jeyu@kernel.org
Cc: torvalds@linux-foundation.org
Link: https://lkml.kernel.org/r/20180116205228.4890-1-andi@firstfloor.org