x86/idle: Control Indirect Branch Speculation in idle
Indirect Branch Speculation (IBS) is controlled per physical core. If
one thread disables it then it's disabled for the core. If a thread
enters idle it makes sense to reenable IBS so the sibling thread can run
with full speculation enabled in user space.

This makes only sense in mwait_idle_with_hints() because mwait_idle()
can serve an interrupt immediately before speculation can be stopped
again. SKL which requires IBRS should use mwait_idle_with_hints() so
this is a non issue and in the worst case a missed optimization.

Originally-by: Tim Chen <tim.c.chen@linux.intel.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>