HID: i2c-hid: Do not bind to CHPN0001 touchscreen
 The CHPN0001 ACPI device has a _CID of PNP0C50 but is not HID
compatible,it uses its own protocol which is handled by the
chipone_icn8318 driver.

If the i2c_hid_driver's probe functon gets called it will fail with a
"hid_descr_cmd failed" error.

Worse, after the probe failure the i2c / ACPI core code will put the
ACPI device in D3 state and when the chipone_icn8318 driver then loads
the device is put back in D0 state, executing its PS0 ACPI method, which
resets the controller, causing the controller to loose its firmware
which was loaded by the BIOS. The chipone_icn8318 driver has a
workaround for this, but that requires it to be the only (or the first)
driver to probe the device.

This commit adds a match callback and returns -ENODEV for i2c_client-s
with a CHPN0001 ACPI device id, so that the probe function never gets
called, fixing the controller losing its firmware.

 Signed-off-by: Hans de Goede <hdegoede@redhat.com>
 Reviewed-By: Benjamin Tissoires <benjamin.tissoires@redhat.com>
 Reviewed-by: Jiri Kosina <jkosina@suse.cz>