Revert "x86/retpoline: Simplify vmexit_fill_RSB()"
 By putting the RSB filling out of line and calling it, we waste one RSB
slot for returning from the function itself, which means one fewer
actual function call we can make if we're doing the Skylake abomination
of call-depth counting.

It also changed the number of RSB stuffings we do on vmexit from 32,
which was correct, to 16. Let's just stop with the bikeshedding; it
didn't actually *fix* anything anyway.

Signed-off-by: David Woodhouse <dwmw@amazon.co.uk>